Merging r143805: ------------------------------------------------------------------------ r143805 | chandlerc | 2011-11-05 03:15:30 -0700 (Sat, 05 Nov 2011) | 4 lines
Switch these two tests to use the Clang driver instead of CC1. They want
to do "realistic" includes, and so need the header search logic now in
the driver. This in turn requires switching the CC1 options to the
actual driver options, and passing -Xclang where there is no analogy.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project@143934 91177308-0d34-0410-b5e6-96231b3b80d8