Merging r143840: ------------------------------------------------------------------------ r143840 | chandlerc | 2011-11-05 15:23:14 -0700 (Sat, 05 Nov 2011) | 4 lines
Remove a pointless member. I have no idea why I made this not a local
variable to begin with... As I'm planning to add include root
information to this object, this would have caused confusion. It didn't
even *actually* hold the include root by the time we were done with it.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project@143942 91177308-0d34-0410-b5e6-96231b3b80d8