Merging r143874: ------------------------------------------------------------------------ r143874 | chandlerc | 2011-11-06 02:31:01 -0800 (Sun, 06 Nov 2011) | 16 lines
Throw the switch and move all Linux header search over to the GCC
detection system that is providing the library paths and crt object
files.

This, modulo any bugs that need to be shaken out, resolves numerous bugs
with how we handle header paths. Here are a few that I know of:
- We no longer need to enumerate all GCC versions searched.
- OpenSUSE searched GCC versions in the wrong order.
- There were typos when selecting various patterns, etc.
- We aren't stating quite some many directories now.
- SysRoot didn't always work in a reasonable way.

I'm working on tests for this, but the tests are making me and Lit sad.
The real testing for this type of driver change is to try it out on
various distributions. I'll hit the common ones right away, and start
more thorough testing tomorrow after some sleep.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project@143951 91177308-0d34-0410-b5e6-96231b3b80d8