Code drop of the current state of the tooling and AST matcher work at Google. Most of this code is authored by Manuel Klimek, some of the early code was originally authored by Zhanyong Wan, and only style tweaks and build system munging from me.
All of this is slated for incremental review and migration to the primary Clang
codebase in one form or another. We're putting the code here early so that
others can see it and gain some context on what we're doing.

To use any of this, you need both the JSONParser that is currently in review
and the VariadicFunction that is currently in review. Both of those are slated
for going into LLVM, not Clang.


git-svn-id: https://llvm.org/svn/llvm-project@146652 91177308-0d34-0410-b5e6-96231b3b80d8