Reenable building with -fvisibility-inlines-hidden.
This was disabled years ago because of a bug in GCC 4.1, which is
on our "broken compilers" list for other reasons. Saving ~500k
on a clang binary (Release+Asserts) is well worth dropping support
for it.

We currently disable it for shared libraries (where it would bring
the biggest win) because clang is broken (PR11642).

IMPORTANT: If you're doing incremental builds you may get tons of
linker warnings. make clean will fix them.

git-svn-id: https://llvm.org/svn/llvm-project@147182 91177308-0d34-0410-b5e6-96231b3b80d8