<rdar://problem/10760649>
Fixed an issue with closing debug sessions down where we might end up
double closing a file descriptor. Previously we had a mutex in a class
that was being used in a read thread, and the main thread would come along
and cancel the thread during the file close and the thread would exit without
releasing the mutex and cause deadlocks. I solved the issue by not canceling
the read thread and closing all needed file descriptors.



git-svn-id: https://llvm.org/svn/llvm-project@149346 91177308-0d34-0410-b5e6-96231b3b80d8