No actual functional change here, just some clarifications:
Clarify that contributors are agreeing to license their code under the 
license corresponding to the subproject that they are contributing to,
as requested by a potential contributor.

Also, as a drive-by, update the llvm-gcc/GPL section to mention dragonegg
and say that GPL code is all in its own repo's.



git-svn-id: https://llvm.org/svn/llvm-project@150065 91177308-0d34-0410-b5e6-96231b3b80d8