<rdar://problem/11072382>
Fixed a case where the source path remappings on the module were too expensive to
use when we try to verify (stat the file system) that the remapped path points to
a valid file. Now we will use the lldb_private::Module path remappings (if any) when
parsing the debug info without verifying that the paths exist so we don't slow down
line table parsing speeds.



git-svn-id: https://llvm.org/svn/llvm-project@153059 91177308-0d34-0410-b5e6-96231b3b80d8