Switches the JSONCompilationDatabase to use the YAML parser. This will allow us to delete the JSON parser from llvm.
The biggest change is a general change of strategy - instead
of storing StringRef's to the values for the command line and
directory in the input buffer, we store ScalarNode*'s. The
reason is that the YAML parser's getRawValue on ScalarNodes
returns a string that includes the quotes in case of double
quoted strings.

For the same reason we're removing the JSON parsing part of
the command line parsing - this means an extra copy for a
command line when it is requested (and only when it is requested).



git-svn-id: https://llvm.org/svn/llvm-project@154929 91177308-0d34-0410-b5e6-96231b3b80d8