<rdar://problem/11282938>
Fixed an issue where we get NULL compile units back from the symbol vendor. We need symbol vendors to be able to quickly give an estimate of the compile units that they have without having to fully vette them first, so anyone getting compile units from a module should be able to deal with a NULL compile unit being returned for a given index.



git-svn-id: https://llvm.org/svn/llvm-project@155398 91177308-0d34-0410-b5e6-96231b3b80d8