Update SmallVector to support move semantics if the host does. Note that support for rvalue references does not imply support for the full set of move-related STL operations.
I've preserved support for an odd little thing in insert() where
we're trying to support inserting a new element from an existing
one.  If we actually want to support that, there's a lot more we
need to do:  insert can call either grow or push_back, neither of
which is safe against this particular use pattern.

git-svn-id: https://llvm.org/svn/llvm-project@155979 91177308-0d34-0410-b5e6-96231b3b80d8