Clean up the usage of "MasterPlan" status in ThreadPlans.  Only user-initiated plans should be MasterPlans that want to stay on the plan stack.  So make all plans NOT MasterPlans by default and then have the SB API's and the CommandObjectThread step commands set this explicitly.
Also added a "clean up" phase to the Thread::ShouldStop so that if plans get stranded
on the stack, we can remove them.  This is done by adding an IsPlanStale method to the
thread plans, and if the plan can know that it is no longer relevant, it returns true,
and the plan and its sub-plans will get discarded.


git-svn-id: https://llvm.org/svn/llvm-project@156101 91177308-0d34-0410-b5e6-96231b3b80d8