test/Makefile: only run lit tests by default, instead of try all of test-suite
Change the "test" target that used to run poolalloc on everything in test-suite
to "fulltest", preserving this functionality should it be used/desired.

It might make sense to still run some of test-suite by default (Olden?),
  but I'm deferring policy decision on that for now.

As far as I know no one developing DSA runs 'make' or 'make test' in this
  directory anyway, so we're only hurting new users by this complicated and
  overkill default testing target (not to mention it's another dependency).

(see comments on PR12744)

This also better fits what we actually support/test frequently :).

git-svn-id: https://llvm.org/svn/llvm-project@156516 91177308-0d34-0410-b5e6-96231b3b80d8