<rdar://problem/11388521>
Fixed a case where the install header phase in the Xcode project would not set the LLDB_VERSION #define correctly. It has now been fixed to key off of the CURRENT_PROJECT_VERSION which will get auto updated with "agvtool bump -all". This was done by adding a build setting to the install-headers makefile based target.



git-svn-id: https://llvm.org/svn/llvm-project@156535 91177308-0d34-0410-b5e6-96231b3b80d8