<rdar://problem/11455398>
Add "--name" option to "image lookup" that will search both functions and symbols.

Also made all of the output from any of the "image lookup" commands be the same regardless of the lookup type (function name, symbol name, func or symbol, file and line, address, etc). The --verbose or -v option also will expand the results as needed and display things so they look the same.



git-svn-id: https://llvm.org/svn/llvm-project@156835 91177308-0d34-0410-b5e6-96231b3b80d8